Crate esp_riscv_rt

Minimal startup / runtime for RISC-V CPUs from Espressif

Features

This crate provides:

Before main initialization of the .bss and .data sections controlled by features

#[entry] to declare the entry point of the program

Re-exports

pub use self::Interrupt as interrupt;

pub use riscv;

Structs

TrapFrame

Registers saved in trap handler

Functions

start_rust⚠

Rust entry point (_start_rust)

start_trap_rust⚠

Trap entry point rust (_start_trap_rust)

Attribute Macros

entry

Attribute to declare the entry point of the program

pre_init

Attribute to mark which function will be called at the beginning of the reset handler.